refactor: minor fixes
- fix calculating of serializer writer
- add writing size while serializing slice
- remove most of deadcode

+++ src/serializer/nfs/read.rs
@@ -4,7 +4,7 @@ use std::io;
 use std::io::Write;
 
 use crate::serializer::nfs::files::file_attr;
-use crate::serializer::{bool, option, u64};
+use crate::serializer::{bool, option, u32};
 use crate::vfs::read;
 
 /// Serializes the non-payload part of [`read::SuccessPartial`] (READ3resok body) into XDR.
@@ -12,7 +12,7 @@ use crate::vfs::read;
 /// The actual read data bytes are sent separately.
 pub fn result_ok_part(dest: &mut impl Write, arg: read::SuccessPartial) -> io::Result<()> {
     option(dest, arg.file_attr, |attr, dest| file_attr(dest, &attr))?;
-    u64(dest, arg.count)?;
+    u32(dest, arg.count)?;
     bool(dest, arg.eof)
 }
